refactor: move project creation logic to lib directory
Refactor the CLI tool by extracting the project creation logic into a separate module in the lib directory. This improves code maintainability and reusability.
Additionally, update package.json to include the lib directory in the published files and set the type to "module" to support ES modules.
