+++ .gitignore
@@ -4,7 +4,6 @@
 # Build output
 dist/
 build/
-lib/
 
 # Coverage directory
 coverage/
