+++ lib/index.js
@@ -1,0 +1,1 @@
+export { default as createProject}  from './creata-project.js'
