+++ package.json
@@ -3,11 +3,13 @@
   "version": "0.1.3",
   "description": "CLI tool to create PocketFlow projects",
   "main": "index.js",
+  "type": "module",
   "bin": {
     "create-pocketflow": "./bin/cli.js"
   },
   "files": [
     "bin",
+    "lib",
     "template"
   ],
   "scripts": {
@@ -26,5 +28,8 @@
     "commander": "^9.4.1",
     "fs-extra": "^11.1.1",
     "inquirer": "^8.2.5"
+  },
+  "engines": {
+    "node": ">=18.0.0"
   }
 }
